+++ freelancer/utils/actions.go
@@ -3,21 +3,22 @@ package utils
 // TODO: check this service functionality so the projectActionBody should have better fields
 // check : https://developers.freelancer.com/docs/projects/projects#projects-put
 // Performs an action on a project.
-type ProjectActionBody struct {
+type ActionProject struct {
 	ProjectID int64         `json:"project_id"`
 	Action    ProjectAction `json:"action"`
 }
 
-type ActionProjectCollaborationsBody struct {
-	// required
-	Action ProjectCollaborationAction `json:"action"`
-	// required
-	Permissions struct {
-		Chat     bool `json:"CHAT"`
-		BidAward bool `json:"BID_AWARD"`
-	}
+// action an permissions are required
+type ActionCollaboration struct {
+	Action      CollaborationAction `json:"action"`
+	Permissions Permissions         `json:"permissions"`
 }
 
-type ActionBidBody struct {
+type Permissions struct {
+	Chat     bool `json:"CHAT"`
+	BidAward bool `json:"BID_AWARD"`
+}
+
+type ActionBid struct {
 	Action BidAction `json:"action"`
 }